renamed to minirig
