+++ pacman/minirig/routing_table/remove_default_routes.py
@@ -1,5 +1,5 @@
-from pacman.myrig.routing_table.exceptions import MinimisationFailedError
-from pacman.myrig.routing_table.utils import intersect
+from pacman.minirig.routing_table.exceptions import MinimisationFailedError
+from pacman.minirig.routing_table.utils import intersect
 
 
 def minimise(table, target_length, check_for_aliases=True):
